Move SPDX Licences to a separate field validator and the Licence Object into a generic OAS validator
