+++ amati/fields/spdx_licences.py
@@ -5,27 +5,17 @@ Validates the Open API Specification licence object - §4.8.4.
 import json
 import pathlib
 from typing import Annotated, Optional
-from typing_extensions import Self
 
-from pydantic import AfterValidator, AnyUrl, Field, model_validator
+from pydantic import AfterValidator, AnyUrl
 
-from amati.validators import title
 from amati.logging import Log, LogMixin
-from amati.validators.generic import GenericObject
 from amati.validators.reference_object import Reference, ReferenceModel
 
-oas_reference: Reference = ReferenceModel(
-    title=title,
-    url='https://spec.openapis.org/oas/v3.1.1.html#license-object',
-    section='License Object'
-)
 
 spdx_reference: Reference = ReferenceModel(
     title='SPDX License List',
     url='https://spdx.org/licenses/',
 )
-
-references: Reference = [oas_reference, spdx_reference]
 
 
 DATA_DIRECTORY = pathlib.Path(__file__).parent.parent.resolve() / 'data'
@@ -94,44 +84,3 @@ SPDXURL = Annotated[
 ]
 
 
-class LicenceObject(GenericObject):
-    """
-    A model representing the Open API Specification licence object
-     
-    OAS uses the SPDX licence list
-
-    Args:
-        name: The name of the licence
-        identifier: The SPDX identifier of the licence
-        url: The URL associated with the licence
-    """
-
-    name: str = Field(min_length=1)
-    # What difference does Optional make here?
-    identifier: SPDXIdentifier = None
-    url: SPDXURL = None
-    _reference: Reference = references # type: ignore
-
-    @model_validator(mode='after')
-    def check_url_associated_with_identifier(self: Self) -> Self:
-        """
-        Validate that the URL matches the provided licence identifier.
-
-        This validator checks if the URL is listed among the known URLs for the 
-        specified licence identifier.
-
-        Returns:
-            The validated licence object
-        """
-        if self.url is None:
-            return self
-
-        # Checked in the type AfterValidator, not necessary to raise a warning here.
-        # only done to avoid an unnecessary KeyError
-        if self.identifier not in VALID_LICENCES:
-            return self
-
-        if str(self.url) not in VALID_LICENCES[self.identifier]:
-            LogMixin.log(Log(f'{self.url} is not associated with the identifier {self.identifier}', Warning, self._reference))
-
-        return self
